backend/requirements.txt
fastapi[standard]==0.114.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
aiofiles==24.1.0
aiosqlite==0.20.0
cryptography==43.0.3
pydantic-settings==2.5.2
Faker==30.3.0
PyJWT==2.9.0
openai==1.50.2
python-dotenv==1.0.1
